templates/conversation_manager/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "strands-conversation-yourname"  # TODO: Replace with your conversation manager name
version = "0.1.0"
description = "Your conversation manager description"  # TODO: Add a brief description
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}  # TODO: Add your information
]
keywords = ["strands", "agents", "ai", "conversation", "context"]  # TODO: Add relevant keywords
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "strands-agents>=1.0.0",
    # TODO: Add any additional dependencies your conversation manager needs
    # "tiktoken>=0.5.0",  # Example for token counting
]

[project.urls]
Homepage = "https://github.com/yourusername/strands-conversation-yourname"  # TODO: Update URLs
Documentation = "https://github.com/yourusername/strands-conversation-yourname#readme"
Repository = "https://github.com/yourusername/strands-conversation-yourname"
Issues = "https://github.com/yourusername/strands-conversation-yourname/issues"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0,<9.0.0",
    "ruff>=0.13.0,<0.14.0",
    "mypy>=0.981,<1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/strands_conversation_yourname"]

[tool.ruff]
line-length = 120
include = ["src/**/*.py", "tests/**/*.py"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
